LayoutTests/imported/w3c/web-platform-tests/css/css-color-adjust/rendering/dark-color-scheme/color-scheme-iframe-dynamic-expected.txt



PASS Initially, the preferred color-scheme is the same in both documents
FAIL Modify color-scheme <meta> assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 255, 0)"
FAIL Modify iframe color-scheme assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 255, 0)"
